Change type and purpose of SSL variables
Add kibana_ssl_cert_file and kibana_ssl_key_file variables to README. These variables are for the public and private key files names. If they are not defined, ansible_fqdn is used. I added this behavior to the task so it will match the ssl.inc template.
Remove elasticsearch role dependency.